Storage/S3: Split plugin and storage
